@@ -12,24 +12,26 @@
             "email": "llpereiras@gmail.com"
         }
     ],
-    "minimum-stability": "dev",
     "require": {
-        "php": ">=5.6.0",
+        "php": ">=7.0.0",
         "mikey179/vfsStream": "1.*",
         "dompdf/dompdf": "0.6.*",
         "phenx/php-font-lib": "0.2.*",
         "zendframework/zend-servicemanager": "~2.0",
         "zendframework/zend-barcode": "^2.3",
-        "satooshi/php-coveralls": "dev-master",
+        "satooshi/php-coveralls": "^1.0",
         "smarty/smarty": "~3.1",
-        "nfephp-org/sped-nfe": "4.1.0.x-dev",
-        "phpunit/phpunit": "~5.7",
-        "squizlabs/php_codesniffer": "^3.0@dev"
+        "nfephp-org/sped-nfe": "~5.0",
+        "phpunit/phpunit": "~6.4",
+        "squizlabs/php_codesniffer": "^3.1",
+        "doctrine/instantiator": "~1.0.5"
     },
     "autoload": {
         "psr-4": {
             "Sped\\Gnre\\": "lib/Sped/Gnre/",
             "Sped\\Gnre\\Test\\": "testes/"
         }
-    }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
