@@ -2,12 +2,13 @@
 
 namespace Sped\Gnre\Test\Render;
 
+use PHPUnit\Framework\TestCase;
 use Sped\Gnre\Render\Html;
 
 /**
  * @covers \Sped\Gnre\Render\Html
  */
-class HtmlTest extends \PHPUnit_Framework_TestCase
+class HtmlTest extends TestCase
 {
 
     public function testDeveRetornarUmInstanciaDoBarCode()
