@@ -2,12 +2,13 @@
 
 namespace Sped\Gnre\Test\Sefaz;
 
+use PHPUnit\Framework\TestCase;
 use Sped\Gnre\Sefaz\boolen;
 
 /**
  * @covers Sped\Gnre\Sefaz\ConsultaConfigUf
  */
-class ConsultaConfigUfTest extends \PHPUnit_Framework_TestCase
+class ConsultaConfigUfTest extends TestCase
 {
 
     public function testDeveDefinirAreceitaParaSerUtilizada()
